requirements-dev.txt
pytest
requests
requests_mock
Flask
Werkzeug

Flask
google-generativeai
https://github.com/googlecodex/voeg-gemini-cli-toe-voor-ai-functies
